Refactor|World|Blockmap: BlockmapCellBlock is now inclusive-exclusive
Previously an inclusive-inclusive representation was used. Although
this representation is immediately easier to use when translating a
map space AABB, it means one has to be aware of how this translates
to a blockmap traversal.

A better approach is an inclusive-exclusive range for traversal and
account for this when converting an AABB to a CellBlock.

Also began the process of unraveling some of the more complicated
Blockmap traversal logics ahead of replacing this API with a better,
iterator based design.