API|Map: Revised Intercept representation
One can now reference the Interceptor from which an Intercept was
produced via the Intercept info if needed, simplifying the API. Note
that the context data pointer is still passed as a callback argument
for user convenience.

Also, Interceptor now uses a different internal representation for
intercepts (making this available as-is places unwanted restrictions
on how the engine can manage this info internally).