libgui|GLState: Specifying a subregion for drawing
Setting the "active rectangle" of GLState will cause viewport and
scissor to be transformed into that rectangle as if the active
rectangle covered the entire render target area.

The purpose is to allow rendering the entire view into a subregion
of the window while using viewport/scissor defined for the full
render target.

Currently the active rectangle affects *all* states in the stack.
It would be better to incorporate it into the GLState attributes.