libcommon|libdoom: Running player thinkers and their mobjs in fractional time
Proof of concept where players and their mobjs live in fractional
time while the rest of the world uses only sharp ticks.

The duration of the current game tick is stored in an engine-side
global variable. The mobj thinkers and playsim physics can access
it using the DD_GAME_TICKF() macro.

This is a work in progress as only libdoom has been updated so far.

Todo:
- gravity not fractional
- rising over steps not fractional
- handling momentum during wall sliding is incorrect
- updating other game plugins except libdoom