Renderer: Disabled old code for halo/flare rendering completely
The objective of this branch is to recreate halo/flare rendering
using GL2.

After that is done, we should evaluate how to integrate it into the
resource management system (or how the resource management system
can co-operate with libgui; probably mostly at the level of locating
the right images to use for the flares).

The terminology should be clarified: a "halo" is the lit 3D volume
around a light source. It is most visible in fog. A "flare" is a
lens flare occurring inside the camera lens(es).