Client: Added CompositorWidget, testing it in ClientWindow
CompositorWidget uses a texture target for all its children.

ClientWindow is now set up to use a compositor for everything that
is not the game view. This should only be done when necessary, not
always as currently.

GuiRootWidget::addOnTop() was added so that widgets can be added
easily into the compositor or the root if a compositor is not in use.