GL|CompositorWidget: Adjusting blending for offscreen composition
The crucial thing is how the alpha channel information is written
in the render target. Normally this can be ignored, however since
the offscreen target then needs to be blended back to the regular
frame, the destination alpha must be valid.

Now using a blending function where the alpha channel is always
additive. This produces results that are closer to non-composited
blending, however there is still room for improvement.