Busy Mode|API|libcommon: Busy transition frame is rendered rather than grabbed
This commit changes how the “screen capture” is done for busy mode.
Instead of taking a screen shot of everything in the window (which
may contain elements that should not freeze, or may be affected by
a VR transformation), the Game and GameUI widgets are drawn to a
texture.

However, at the time of the busy mode start, the game state
may already have been modified so that redrawing a frame would not
result in the expected “frozen” view. To get around this, at certain
times like just before executing a game action (e.g., start new game),
the transition frame is always rendered, in case a busy mode will
need it.

Todo: Some slight glitches remain…