Fixed|X11: Avoid GL error when starting in maximized window
When starting with a maximized window, X11 sends two window resize
events in a rapid succession. Perhaps the changes from the first
resize still haven't resolved fully on the X11 server when the second
one is applied, and fails?

Now window resizes are slightly deferred under X11 so that
redundancies can be ignored.