Performance|Renderer|Refactor: Retain constness in APIs; avoid redundant memory allocs
It is important not to accidentally lost constness of object
references, particularly when dealing with Qt containers. Non-const
access to containers may force Qt to detach the data into a separate
private instance.

However, the most impactful change is the removal of the QBitArray
in Rend_RadioBspLeafEdges(). Allocating and freeing the array in
complex maps is much slower than simply calculating dot products
for the planes as we go.