Fixed|UI|All Games: Transition after selection a menu item
When a new game is started from the Main Menu, the menu is closed
before the map is loaded. Before, this was masked by the fact that
a frame wasn’t drawn before the busy mode was started. Now, the
transition frame is drawn after the menu is closed and the true
behavior becomes visible.

If we actually do want to “freeze” the view with the menu open, we
shouldn’t close the menu before starting to load the next map.