Fixed|ResourceSystem: Automatically destroy a Material along with it's MaterialManifest
Todo: The way this is set up is now somewhat confused. The original
idea was to allow multiple material URIs to resolve to one Material.
However because a Material depends on the manifest for some of it's
configuration info the latter is no longer performing the same role
it used to have, prior to the switch to automatical generation of
material definitions.