Fixed|Sector: Ensure sector lightlevels are clamped to the normalized [0..1] range
Note that presently Doomsday does not support id Tech 1 map hacks
which depend on sector lightlevels outside of the [0..255] range that
vanilla expects.

(Information on what the mod author actually intends with such hacks
appears to be almost non-existent on the web -- please do let us know
what these are actually used for so that we can look to support them
in future).

IssueID #1668