Renderer|libdoom: Use postfx for invulnerability effect
The “post.fx.monochrome.inverted” shader is used as the
invulnerability effect in libdoom.

The blinking effect when the power-up is going away now conforms to
vanilla in that it toggles between completely on and off.

The “postfx” command was enhanced to allow specifying a manual
opacity factor for the effect.

IssueID #1685