UI|Style: Added “-fontsize” option to change font sizes
This is a temporary solution to the Rift UI font sizing. One needs
to manually specify a suitable font size (e.g., “-fontsize 1.5”).

The correct way would be to create an alternative UI style pack for
the Rift and allow switching UI styles on the fly. That will need
some additional work, though, as all widgets are not prepared for
dynamic UI style changes.