Fixed|Stereo 3D|libgui: Visual discrepancies between left/right eyes
When a target was cleared, the current GLState wasn’t being
automatically applied. This meant that the right eye’s UI compositor
target wouldn’t be cleared after the left side had been drawn.

Also, the size of the UI compositor should note the active rectangle
so that the right width is used for the composition.