Fixed|Menu|libcommon: Tiny menu cursor with some mods (e.g., Hell Revealed II)
The size of the menu cursor is determined dynamically according to
the dimensions of the currently focused menu item.

Some mods prefer to use a single image for a menu page background
(typically, the M_DOOM patch) and replace all the M_NGAME (etc...)
patches with (perhaps tiny) images with no color information.

In this situation, the default menu cursor size should be at least
equal to the effective line height of the page.

IssueID #1133