Oculus Rift|Stereo 3D: Improved latency of head tracking
This commit improves the integration of Oculus Rift head tracking
into player view angles.

For the best experience, it is imperative that no kind of additional
latency is introduced to the head tracker’s angles. The renderer
will now use the latest head angles as-is, added to the regular
key/mouse controlled angles.

To make it possible for the head tracker angles to affect both the
renderer and gameplay, the “body yaw” angle is now removed from the
angles reported from the game to the engine, so that the latest
yaw angle may be used instead. (The events of the input subsystem
and the game ticks have some inherent latency that must be avoided.)

Also replaced the VR view angle holding functions with explicitly
controlled head tracker updates. There’s at most one update per
frame, but it’s done at the latest time possible.