Fixed|Particles: Particle spawn position affected by source float bobbing
When particles are spawned, in addition to floorclip, float-bobbing
is also taken into account.

Todo for later: It would be useful to have a generator flag that
determines that bobbing should be applied at a later stage, e.g.,
during rendering. This would allow for particles that appear
inside an object to stay inside it. In practice, it would be enough
to store the “bobbing phase” of the particle and have the renderer
figure out the offset based on that and time.

IssueID #932