UI|Client: Crash when deleting a menu widget
The ownership of menu items’ sub-widgets is dual: as popups, they
always have a parent and if the parent is deleted, the sub-widget is
deleted, too. However, when the data item is deleted, the sub-widget
also needs to be deleted.

Now the data model is explicitly cleared before a menu is deleted,
so that possible sub-widgets are deleted at the right time.