Sky|Added: Support scrolling skies vertically
The sky renderer now supports scrolling layers on both axes. However,
the fadeout colors are not currently updated and will continue to use
the averaged color of the top and bottom lines of the texture regardless
of the current vertical offset.

While it may be possible to sample the fadeout colors dynamically once
rendering of the sky is revised, given that a vertically scrolling layer
should use a seamlessly repeating texture in any case, it is probably
not necessary.

The new definition property Sky.Layer.Origin was added to allow both
axes to be set at once. The existing Sky.Layer.Offset property is not
affected.