Refactor|Console|API: Printing log messages, removed old Console API functions
The engine should not internally use Con_Message, Con_Printf, or
Con_FPrintf to print log messages, as these do not provide the means
to indicate appropriate metadata for the entry.

Con_Printf and Con_FPrintf were removed altogether. All use of them
was replaced with LogBuffer_Printf.

Todo: Apply metadata to internal entries and in the plugins.