Fixed|InFine: Animation gets stuck (suspended) after playdemo
This commit fixes the issue where the game view would be obscured
by the title animation after a game was started. The problem was
that the “playdemo” command returned true even though the demo
was not started (to avoid a warning about the command failing).
This meant the finale was indefinitely suspended.

Demos are not supported at the moment, so disabled the InFine
playdemo instruction (which also avoids the console warning).