Fixed|Hexen|SndInfoParser: SIGSEGV attempting to parse a SNDINFO lump intended for ZDoom
An error is thrown if a lump presented to SndInfoParser() contains
unknown commands.

The vanilla parser attempts to ignore any unknown commands, however,
due to the primitive syntax it is not possible to reliably determine
the end of said command. Consequently attempting to skip the command
may lead the parser into further trouble. As such, it is better to
simply trigger an error and don't bother trying to skip.