Fixed|Oculus Rift|libappfw: Shutting down LibOVR (dangling references)
One of the references to the created LibOVR objects was not freed, thus
causing shutdown of de::OculusRift to hang waiting for threads to stop.

Also, now de::OculusRift initialization is attempted once, after which
no further attempts will be made until deinit() is called. This avoids
retrying the initialization every time de::OculusRift is accessed.