Fixed|libcommon|Hexen DK: Crash during hexen-dk game init (invalid nextMap)
Until a game has begun (and a map has been loaded) there is no valid
value for the current or next map. In previous versions the default
zero-value would resolve to the "default mapinfo" definition. These
days there is no default mapinfo in Hexen, leading to the crash.

For now we'll simply allow the invalid values to resolve to a zero
length URI, meaning no mapinfo will be returned. So, test for this
when initializing a finale animation's state condition flags.