libcommon: Removed cvar 'game-fastmonsters'
Enabling/disabling of fast monsters and/or projectiles must be done
via the game rule mechanism. Game rules are applied on server side
and communicated to the client when changed.

Game rule set management will soon be done through a new UI. As this
cvar has never worked correctly, we might as well dump it.