libgui: Added swap control for X11, debugging Canvas recreation
GLShader, GLProgram, and GLFramebuffer now support recompilation/
has any effect on the OpenGL errors occurring after switching the Canvas
(it didn't).

GuiApp has an audience that gets notified when the Canvas is recreated.

Added swap control for X11 using the GLX function for this purpose so
that Canvas recreation is not necessary when changing vsync on X11
(which was causing OpenGL errors).