libcommon|SaveInfo: Saved games now record the game identity key (was gamemode)
If SaveInfo is to become game-agnostic then we need to refer to games
uniquely with identifiers which are known to the engine. The obvious
choice (and most future proof) is to use the unique identity key for
this purpose.

Gamemode identifiers in old save games are automatically converted in
the process of interpreting the game session header.

As the save version number has already been bumped to v14 for the
Doomsday 1.14 release, saved games made with an unstable build from
build 1131 onward will NOT be loadable.

Also applied the pimpl idiom to SaveInfo, plus further cleanup.