Fixed|libcommon|Game Menu: Error at startup and further compiler warning fixes
On second thought, replacing memory management in the crusty old game
menu, with an OO model, is simply not feasible given the complete lack
of a logical build up and tear down structure to this code.

I think it best to simply work with/around it until such time as the
whole module can be replaced in one fell swoop. There is nothing to
be gained from cleaning this up.

Now to test a build with something other than MSVC...