libcommon|SaveInfo|SaveSlots: Associate the saved game state file name with SaveInfo
When a new SaveInfo is instantiated or copied from another save slot
the game state file name bound to associated save slot is applied.
This file name can be directly queried from the SaveInfo and thus
avoiding the need to pass it along with said info, or, interact with
the SaveSlots container/wrapper.