UI|Client: Added a rudimentary manual "Connect to Server" dialog
This dialog is basically a GUI for the 'connect' console command.
It is only available when a game is loaded, as 'connect' cannot
change the current game.

A more full-featured dialog would first query the given address and
then display some indication of what has been found, if any.

A drop-down list of known server addresses might also be handy
(cf. Shell).