libdeng2|App: Using a .pack as basedir; configurable config path
This commit adds flexibility to App. It is now possible to point the
application's basedir to a pack file, so that the contents of the pack
are used to feed the App's root folder. In this case, the usual /data
and /modules are expected to be found inside the pack.

Also, one can now specify a different path where Config reads its script
from (defaults to "/modules/Config.de" as before).