UI|Multiplayer: Improved manual connection dialog
The manual "Connect to Server" dialog is now substantially more
sophisticated. Once an address has been entered, it starts a query
using ServerLink and shows the results in an MPSelectionWidget.
The game can then be joined by pressing Enter, clicking the dialog's
Connect button, or clicking the MPSelectionWidget's join button.

The dialog takes advantage of an acceptance action to execute the
join. This also works around the issue of not being able to change
games when joining with the "connect" console command.