libgui|GLBuffer: Added "TexBounds0" vertex attribute
The texture bounds attribute is useful for example when rendering
geometry that uses a shared texture atlas; one can store the individual
texture bounds in the vertex data and use normalized (0,1) UV
coordinates, which a shader will then convert to coordinates on the
atlas. This also allows repeating UV patterns by applying wrapping in
the shader.