libcommon|MapStateReader: Replaced map state data Reader
Avoiding the need to update all of the existing deserialization logic
the new reader takes advantage of the callback nature of libdeng1's
Reader, using this is a C API which wraps an instance of de::Reader
that takes the source de::File as input.

As of this commit, libcommon is no longer dependent on the external
LZSS library for savegame deserialization.

Todo for later: Update all map state deserialization logic to use the
native serialization components provided by libdeng2.