Refactor: Continued splitting up remaining old savegame system functionality
It is no longer necessary for the engine to be made aware of which
map state reader is actually used for deserialization. This logic
can exist entirely in libcommon. In turn, there is no need to link
the sessions to the SavedSessionRepository, which, is now little
more than a collection of resource search routines.