libappfw|GuiRootWidget: Use a TextureBank for common UI textures
While the UI images are still stored in an atlas as before, a Bank
is smarter with managing the allocations: only the needed images
are allocated, and multiple instances of widgets can easily share
the same textures.

For instance ProgressWidget was a big offender in allocating the
wheel/gear textures separately for each instance.