libcommon|MapStateWriter: Replaced map state data Writer
Avoiding the need to update all of the existing serialization logic
the new writer takes advantage of the callback nature of libdeng1's
Writer, using this as a C API which wraps an instance of de::Writer
that takes the source de::File as input.

As of this commit, libcommon is no longer dependent on the external
LZSS library.

Todo for later: Update all map state serialization logic to use the
native serialization components provided by libdeng2.