libdeng2|SavedSession: Removed "version" from SessionMetadata
This artifact of the old savegame system is no longer used with the
new .save packages as each serialized state file as its own version
number (in the file header).

Old savegames converted using Savegame Tool are marked with the old
save version number (in the Info header) for reference and/or future
debugging purposes.