libdeng2|SavedSession: Use absolute paths to SavedSessions in the "repository" index
Although the user's savegames will always exist in /home/savegames
we should require that all game saving is done here. For example,
Hexen uses a "base" savegame internally for storing the states of
visited maps in a hub for the current game session. This built-in
savegame should not be stored alongside the user's savegames.