libcommon: Disabled manual caching of SavedSession metadata on write
The case sensitivity mismatches between Info and Record currently
result in the parsed metadata not matching that which we try to
manually cache to avoid reopening of the .save package. Although
the values are the same, their names are not.

For now we'll let SavedSession cache it itself when called for.