All Games: Unified game session model
All supported games now use the same logical model for managing the
game session and the associated saved sessions. An internal savegame
is now used for all games (irrespective of the user's savegame prefs)
which is now managed automatically by the GameSession class.

See the GameSession API documentation for full details of the changes.

Also tweaked the domain and verbosity of various log messages.

Todo for later: Cleanup GameSession-internal logic and FS2 usage.