libdeng2|Session: game::Session provides access to the resource file list
game::Session now provides access to the Profile for the current game
session. This profile includes the list of resources files pertinent
to the current (perhaps implicit) game session, as specified on the
command line.

This means that everyone with visibility of game::Session can now access
this information for whatever reason.

In the future Session should store the Profile persistently (perhaps a
specialized SettingsRegister would be good here). SavedSession should
also record this information to assist in determining whether it is
compatible with the current Session::Profile.