libdeng2|FS: Improvements for accessing Archive through file system
ArchiveEntryFile now has a meaningful flush() method that causes the
entire archive to be written to the source file (if modified).

ArchiveFeed write access was clarified: if the source file has write
access at the time of ArchiveFeed creation, the populated files
and subfolders for the archive contents will be created with Write
mode.