Client|ResourceSystem: Invalid sprite frame names should not cause alert
Now the log entries about invalid sprite frames are logged as NOTE,
meaning they will still be highlighted in the log however will not
cause an actual alert.

Addresses the spurious alerts in Heretic and Hexen.

Todo for later: This might be controlled by a blacklist so that known
IWADs would not raise alerts?