Fixed|libcommon: Applied undefined game rules loading legacy savegame
If the game plugin is not unloaded when the game mode changes then
the rules will be left in an undefined state prior to the new game
beginning. Although the game rules are always set when a new session
is begun. However, rather than beginning a new session, if a converted
legacy savegame is loaded which lacks a complete game ruleset then
the undefined game rules from the previous game session were applied
erroneously.

The game rules are explicitly cleared in G_CommonPreInit()