Savegame Tool|NativeTranslator: Extract more metadata from legacy Hexen savegames
The present player list and time in the current map is now extracted
from old Hexen savegames and included in the metadata (the Info file)
similarly to the other games. While Doomsday does not currently use
these with Hexen it is useful info that belongs in the metadata (and
also avoids having to retrieve it from the map state data file).