Console|Resources: Generalized "inspectsavegame", moved into the engine
The inspectsavegame command can now be used to inspect the savegames
of any game at any time (e.g., while in Ring Zero).

The saved session to be inspected is specified by absolute path in the
virtual file system (e.g., /home/savegames/doom1-ultimate/doomsav1).
If the saved session to be inspected belongs to the currently loaded
game then the full path may be omitted.