Fixed|libcommon: ANIMATED lump ZDoom extension incompatibility
ZDoom interprets the first byte in a texture animation as a set of
flags rather than a boolean, adding a 'No Decals' flag into the mix.
Previously a non-zero value was interpreted to mean the animation is
defined in the 'Textures' resource scheme. Consequently resulting in
definitions being added into the wrong scheme.