Homepage|Forums: Responsive reformatting of forum/topic listings
If the width of the window is small enough that a tabular layout for
forum and/or topic listings negatively affects readability, then use
a multi-line layout to visualize the same info.

Note that forum/topic list icons have been disabled temporarily. This
will return once it is clear how they should be incorporated into the
responsive reformatting.