Renderer|BspLeaf|Line|SectorCluster: SectorCluster is responsible for BiasIllum management
SectorCluster is now responsible for the generation of map geometry
"Shards" within the cluster and allocates bias illumination points as
needed when first drawn. "Splitting a Cluster into Shards" has quite
a nice ring to it :)

The now redundant BiasSurface was removed.

At present a Shard consists of the Bias lighting data only. No vertex
geometry is yet stored here.

Todo: The Shard struct should now begin to evolved towards the 2.0
map renderer design for this component. Presently its internal to
SectorCluster as nobody else needs access to it.