Refactor: Added libdoomsday, relocated some files
libdoomsday is the common denominator between the client and the
server applications.

Unlike liblegacy, it is intended to be a permanent and central part
of the architecture. All shared functionality that exists both in
the client and the server (but not at a more general level) should be
relocated there, after appropriate refactoring.

Relocated the following:
- logical sound manager
- old WAV parser
- base mobj struct declaration
- DualString