Cleanup|Client: Removed UI background InFine script
The UI framework is in control of drawing the UI (background included)
so InFine is no longer being used for this purpose.

In the future, once there is Doomsday Script based InFine and an
InfineWidget, the window background image might be replaced with one,
though.

Fixes the warning where the background finale was trying to access the
removed "background" image resource.