Refactor|libdoomsday|DED: Definition memory management
This commit applies some C++ sugar so ease the handling of the DED v1
definitions. Since these are C POD structs, a DEDArray<> template
was added to help with the allocating and releasing of memory owned
by elements. DEDArray also handles the dynamic resizing of the
array, as ded_count_s has been incorporated into it.

Each individual DED data struct is now responsible for releasing and
reallocating memory owned by it (such as Uris). These methods will
be called by DEDArray as needed. The inline copy logic from the DED
parser was moved to the DED structs.