Renderer: Revert to using local WallEdges in writeAllWallSections
On second thoughts, the existing map renderer is simply not ready to
support such a fundamental design change. Until the renderer is able
to draw the world as homogeneous chunks of geometry, internalization
of WallEdge in SectorCluster is pretty much impossible.

Todo:

Revise the map geometry pipeline to work using wall geometries that
have already received "edge divisions" at a much higher level.

Also, deferring the drawable geometry copy into the render list's
internal vertex buffer until it has been fully prepared not only adds
significant and unnecessary overhead but also complicates the whole
pipeline. This is due to the need to drag map data and/or structures
deep into the inter bowels of the renderer. (WallEdge being one such
example of data that does not belong in the guts of the renderer).

Instead, an API should be introduced to reserve storage in the vertex
buffer as soon as possible so that map geometries can be constructed
in-place (for now) in a form that is immediately drawable when done.