Optimize|Renderer|DrawList: Support sequential vertex indices in DrawList
Currently the map renderer will always allocate vertex elements from
the world vertex buffer in a linear sequence. This means that is not
necessary to store the whole sequence in a DrawList.

DrawList can now be configured to derive the buffer indices using the
indice base and range only. As the indice sequence for each element
is usually allocated from the same memory region as the list itself,
using sequential indices avoids the need to resize (and relink the
sequence array pointers) when an element is written to the list.

Also removed the in-list vertex buffer pointer; we can safely assume
that all geometry comes from The One world vertex buffer.