Windows|qmake: Basic configuration for MinGW
Now that Doomsday is based on Qt, support for MinGW is within the
realm of possibility. This commit sets up the basic build rules.
However, a successful build is not yet possible, as many source files
assume WIN32 implies MSVC/_MSC_VER.

Fixed a couple of issues in the Doomsday SDK libs when compiling
with MinGW.