World|ConvexSubspace|Client: ConvexSubspace no longer remembers last sprite projection frame
This info is redundant given that mobj_t remembers it. Also, because
the contact spreading algorithm is used for mobjs to prevent their
popping into view around corners; the subspace they are linked to at
render time may not be the same subspace attributed to the BSP leaf
at their origin.