Fixed|liblegacy|libdoomsday|libcore|Windows: File path encoding issues
This commit fixes a number of issues when handling file paths encoded in
UTF-8.

On Windows, the native file APIs used a different 8-bit encoding; all
use of native file APIs was updated to the wchar variant of the APIs.
Also, fs_util.h now defines macros that redirect fopen, mkdir and access
to the UTF-8 friendly version of these functions.

NativePath would in some methods accidentally apply an ASCII-encoding
when given a 8-bit path (automatic char* -> QString conversion). This
was also fixed; it should have an effect on all platforms, not just Windows.