Refactor|SectorCluster|Client: Reorganized SectorCluster-internal data structures
Separate the cluster internal data by ConvexSubspace, placing it in
a "Subsector" structure (not to be confused with the component of the
same name, in the id Tech 1 map format). As the subspace geometry is
now modeled independently (in terms of the halfedge data structure)
and the subsector properties are used only by SectorCluster there is
no longer a need for a client specialized ConvexSubspace.

Todo: Move client side of ConvexSubspace to SectorCluster::Subsector